Fix oss delete function

+++ src/main/java/oss/fosslight/service/impl/OssServiceImpl.java
@@ -1021,7 +1021,6 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 
 			if (StringUtil.isEmpty(ossId)){
 				ossMaster.setCreator(ossMaster.getLoginUserName());
-				
 			} else {
 				orgMasterInfo = new OssMaster();
 				orgMasterInfo.setOssId(ossId);
@@ -1365,6 +1364,8 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 			ossMaster.setOssName(beforeBean.getOssName());
 		}
 		
+		boolean ossCommonDeleteFlag = false;
+		
 		// 바로 삭제 일 경우( identification 상태가 conf인 프로젝트가 없을시 )
 		if (CoConstDef.FLAG_NO.equals(avoidNull(ossMaster.getNewOssId(), CoConstDef.FLAG_NO))) {
 			//3. 기존의 Oss 삭제		
@@ -1373,6 +1374,7 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 			// 닉네임은 이름으로 매핑되기 때문에, 삭제후에 신규 추가시 자동으로 설정되는 문제가 있음
 			// 삭제하는 oss 이름으로 공유하는 닉네임이 더이상 없을 경우, 닉네임도 삭제하도록 추가
 			if (ossMapper.checkHasAnotherVersion(ossMaster) == 0) {
+				ossCommonDeleteFlag = true;
 				ossMapper.deleteOssNickname(ossMaster);
 			}
 			
@@ -1382,10 +1384,9 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 			
 			updateLicenseDivDetail(ossMaster);
 			
-			ossMapper.deleteOssDownloadLocation(ossMaster);
+			if (ossCommonDeleteFlag) ossMapper.deleteOssDownloadLocation(ossMaster);
 			ossMapper.deleteOssMaster(ossMaster);
-			ossMapper.deleteOssCommonMaster(ossMaster);
-		
+			if (ossCommonDeleteFlag) ossMapper.deleteOssCommonMaster(ossMaster);
 		} else {
 			// 동일한 oss에서 이동하는 경우, nick name을 별도로 등록하지 않음
 			OssMaster afterBean = getOssInfo(ossMaster.getNewOssId(), false);
@@ -1410,6 +1411,7 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 			}
 			
 			if (ossMapper.checkHasAnotherVersion(ossMaster) == 0) {
+				ossCommonDeleteFlag = true;
 				ossMapper.deleteOssNickname(ossMaster);
 			}
 			
@@ -1420,11 +1422,10 @@ public class OssServiceImpl extends CoTopComponent implements OssService {
 			
 			updateLicenseDivDetail(ossMaster);
 			
-			ossMapper.deleteOssDownloadLocation(ossMaster);
+			if (ossCommonDeleteFlag) ossMapper.deleteOssDownloadLocation(ossMaster);
 			ossMapper.deleteOssMaster(ossMaster);
-			ossMapper.deleteOssCommonMaster(ossMaster);
+			if (ossCommonDeleteFlag) ossMapper.deleteOssCommonMaster(ossMaster);
 		}
-	
 	}
 
 	@Override
